Moved scheduleDays.js to helpers and renamed it
