+++ code/TimeManagementApp/helpers/scheduleHelper.js
